controllers create and update improved, some css on nav

+++ app/controllers/schedules_controller.rb
@@ -20,6 +20,9 @@ class SchedulesController < ApplicationController
   # GET /schedules/1/edit
   def edit
     @schedule = Schedule.find(params[:id])
+    unless current_user == @schedule.boat.profile
+      redirect_to schedules_path
+    end
     @today = Date.current
   end
 
@@ -32,23 +35,31 @@ class SchedulesController < ApplicationController
         return
       end
     end
-    if params[:schedule][:start] < params[:schedule][:finish]
-      schedule = Schedule.create(schedule_params)
-      flash[:success] = "Schedule Created"
-      redirect_to schedule
-      return
-    elsif params[:schedule][:start] >= params[:schedule][:finish]
-      flash[:error] = "Schedule Finish Date Must Be After Start Date"
-    else 
-      flash[:error] = "Schedule Unable To Be Created"
+    current_user.profile.boats.each do |b|
+      if b.id == params[:schedule][:boat_id].to_i
+        if params[:schedule][:start] < params[:schedule][:finish]
+          schedule = Schedule.create(schedule_params)
+          flash[:success] = "Schedule Created"
+          redirect_to schedule
+          return
+        elsif params[:schedule][:start] >= params[:schedule][:finish]
+          flash[:error] = "Schedule Finish Date Must Be After Start Date"
+        else 
+          flash[:error] = "Schedule Unable To Be Created"
+        end
+      end
     end
     redirect_back(fallback_location: root_path)
   end
   
   # PATCH/PUT /schedules/1
   def update
+    schedule = Schedule.find(params[:id])
+    unless current_user == schedule.boat.profile
+      redirect_to schedule
+      return
+    end
     if params[:schedule][:start] < params[:schedule][:finish]
-      schedule = Schedule.find(params[:id])
       schedule.update(schedule_params)
       flash[:success] = "Schedule Updated"
       redirect_to schedule
@@ -64,7 +75,12 @@ class SchedulesController < ApplicationController
   # DELETE /schedules/1
   def destroy
     schedule = Schedule.find(params[:id])
-    schedule.destroy
+    if current_user == schedule.boat.profile
+      flash[:notice] = "Schedule Removed"
+      schedule.destroy
+    else
+      flash[:error] = "You are not the boat captain."
+    end
     redirect_to schedules_path
   end
